refactor: separate files in `src\other`
